Run relay-server container as UID 0
The relay-server is hardcoded to bind to port 443, but the binary
doesn't use the `NET_BIND_SERVICE` capability, even if that capability
is available in the container.

To workaround the issue, we grant the vcluster permissions to use an SCC
which has no restrictions on UIDs so that binding to port 443 succeeds.